seprated logic into modules
